made cppcheck compile with g++ 4.3

+++ tokenize.cpp
@@ -8,6 +8,7 @@
 
 #include <map>
 #include <string>
+#include <cstring>
 
 #include <stdlib.h>     // <- strtoul
 #include <stdio.h>
@@ -683,7 +684,7 @@ void SimplifyTokenList()
             free(tok->str);
             char str[10];
             // 'sizeof(type *)' has the same size as 'sizeof(char *)'
-            sprintf( str, "%lu", sizeof(char *));
+            sprintf( str, "%u", (unsigned int)sizeof(char *));
             tok->str = strdup( str );
 
             for (int i = 0; i < 4; i++)
